Save Game, Load Game and Scores
Amongst many minor fixes and changes the three major features of this
patch are Saving Games, Loading Games and Score Storage which is in the
DB for use with the Leaderboard when it is complete. Also worth
mentioning is the finalization of the moving of all assetLoader arrays
to being objects.